Backend: Merge notifications logic into socket.js instead and expose notify() through socket.js
